fix: Hot fixes to staging for composer.space (#6642)
