+++ packages/apps/composer-app/src/beta/notice.md
@@ -1,0 +1,12 @@
+Composer is now in Beta and regular updates will be published to the new site 
+:link[composer.space]{href="https://composer.space"}.
+
+Composer will remain free to use and we are still 100% committed to building privacy-preserving open source software.
+However, we are asking Beta users to sign-up and give us feedback to help us build Composer during this next stage of development.
+
+You will be able to transfer your existing data by initiating a
+:link[device invitation]{href="https://docs.dxos.org/guide/platform/halo.html#device-invitations"}.
+Please click the Sign-up button below to register for the Beta.
+
+If you have questions or comments, please join our :link[Discord]{href="https://discord.gg/eXVfryv3sW"}.
+Thanks for your support!
